fixed MSVC complaint

+++ trunk/src/srgb_render.cc
@@ -62,7 +62,7 @@ cv::Mat Srgb_render::linear_to_sRGB(const cv::Mat& img) {
         csum += histo[minval];
     }
     
-    if (fabs(maxval - minval) < 0.05*(maximum - minimum)) {
+    if (fabs(double(maxval - minval)) < 0.05*(maximum - minimum)) {
         // we are probably looking at a synthetic image with insufficient dark areas to
         // make the 2% threshold meaningful
         // just go with the min/max
